big update changed animation system added deep dream imagery
refactored hacky setTimeout animation system by replacing it with requestAnimation frame

and added new images, that have been run through a deep dream processor
animation system uses ternaries and a if conditional to switch between the animations automatically.

this entails changing the directory string (dirStr)
and the reference to the array of images (picArr) needed for each animation

to delay the timing of the each frame requestAnimation frame is placed in
a setTimeout called every 200ms -- 5fps

to set the interval for the changing background color the timestamp from requestAnimationFrame is passed into someFunc(NEEDS A NEW NAME) and a start time is captured. each time through the setTimeout/fps cycle, a delta is checked between startTime and the current timestamp. if greater than 13000ms changeBkg function is called to transition into the new color by changing the img style attribute